refactor: structural defects phase 2 — ContextVar race, hook dedup, router split (#696)
Phase 7: Fix CRITICAL ContextVar race in invoke_with_timeout() —
  contextvars.copy_context() propagation to ThreadPoolExecutor
Phase 8: Hook deduplication — name-based dedup prevents double registration
Phase 9: LLM router decomposition — adapters.py + provider_dispatch.py
  router.py 1530→1062 lines (-31%)

Quality: ruff 0, mypy 0 (213 files), pytest 3881 pass
