+++ core/llm/router.py
@@ -15,22 +15,27 @@ import os as _os
 import random
 import re
 import time
-from collections.abc import Callable, Iterator
+from collections.abc import Iterator
 from contextvars import ContextVar
 from dataclasses import dataclass
-from typing import Any, Protocol, TypeVar, runtime_checkable
+from typing import Any, TypeVar
 
-import anthropic
 from pydantic import BaseModel
 
 from core.config import (
-    ANTHROPIC_FALLBACK_CHAIN,
-    GLM_FALLBACK_CHAIN,
-    OPENAI_FALLBACK_CHAIN,
     _resolve_provider,
     is_model_allowed,
     settings,
 )
+from core.llm.adapters import CROSS_PROVIDER_FALLBACK as CROSS_PROVIDER_FALLBACK
+from core.llm.adapters import AgenticLLMPort as AgenticLLMPort
+from core.llm.adapters import ClaudeAdapter as ClaudeAdapter
+from core.llm.adapters import LLMClientPort as LLMClientPort
+from core.llm.adapters import LLMJsonCallable as LLMJsonCallable
+from core.llm.adapters import LLMParsedCallable as LLMParsedCallable
+from core.llm.adapters import LLMTextCallable as LLMTextCallable
+from core.llm.adapters import LLMToolCallable as LLMToolCallable
+from core.llm.adapters import resolve_agentic_adapter as resolve_agentic_adapter
 from core.llm.errors import (
     LLMAPIStatusError as LLMAPIStatusError,
 )
@@ -57,6 +62,10 @@ from core.llm.fallback import RETRY_BASE_DELAY as RETRY_BASE_DELAY
 from core.llm.fallback import RETRY_MAX_DELAY as RETRY_MAX_DELAY
 from core.llm.fallback import CircuitBreaker as CircuitBreaker
 from core.llm.fallback import retry_with_backoff_generic as retry_with_backoff_generic
+from core.llm.provider_dispatch import _cross_provider_dispatch as _cross_provider_dispatch
+from core.llm.provider_dispatch import _get_fallback_chain as _get_fallback_chain
+from core.llm.provider_dispatch import _get_provider_client as _get_provider_client
+from core.llm.provider_dispatch import _retry_provider_aware as _retry_provider_aware
 from core.llm.providers.anthropic import (
     FALLBACK_MODELS as FALLBACK_MODELS,
 )
@@ -360,193 +369,6 @@ async def call_with_failover(
         last_error,
     )
     return None, None
-
-
-# ---------------------------------------------------------------------------
-# Provider-aware helpers — route to correct SDK based on model provider
-# ---------------------------------------------------------------------------
-
-
-# ---------------------------------------------------------------------------
-# Provider dispatch — single source of truth for per-provider configurations.
-# Replaces 6 individual _get_provider_*() functions (Kent Beck DRY).
-# ---------------------------------------------------------------------------
-
-# Per-provider circuit breakers (must be module-level singletons)
-_openai_cb = CircuitBreaker()
-_glm_cb = CircuitBreaker()
-
-
-def _openai_retryable() -> tuple[type[Exception], ...]:
-    import openai
-
-    return (openai.RateLimitError, openai.APIConnectionError, openai.InternalServerError)
-
-
-def _openai_bad_request() -> type[Exception]:
-    import openai
-
-    return openai.BadRequestError
-
-
-# Lazy dispatch table — callables to avoid import-time side effects
-_PROVIDER_DISPATCH: dict[str, dict[str, Any]] = {
-    "anthropic": {
-        "get_client": lambda: get_anthropic_client(),
-        "fallback_chain": lambda: list(ANTHROPIC_FALLBACK_CHAIN),
-        "retryable_errors": lambda: _RETRYABLE_ERRORS,
-        "bad_request_error": lambda: anthropic.BadRequestError,
-        "circuit_breaker": lambda: __import__(
-            "core.llm.providers.anthropic", fromlist=["get_circuit_breaker"]
-        ).get_circuit_breaker(),
-    },
-    "openai": {
-        "get_client": lambda: __import__(
-            "core.llm.providers.openai", fromlist=["_get_openai_client"]
-        )._get_openai_client(),
-        "fallback_chain": lambda: list(OPENAI_FALLBACK_CHAIN),
-        "retryable_errors": _openai_retryable,
-        "bad_request_error": _openai_bad_request,
-        "circuit_breaker": lambda: _openai_cb,
-    },
-    "glm": {
-        "get_client": lambda: __import__(
-            "core.llm.providers.glm", fromlist=["_get_glm_client"]
-        )._get_glm_client(),
-        "fallback_chain": lambda: list(GLM_FALLBACK_CHAIN),
-        "retryable_errors": _openai_retryable,  # GLM uses openai SDK
-        "bad_request_error": _openai_bad_request,
-        "circuit_breaker": lambda: _glm_cb,
-    },
-}
-
-
-def _get_provider_config(provider: str, key: str) -> Any:
-    """Lookup a provider-specific configuration by key.
-
-    Valid keys: get_client, fallback_chain, retryable_errors,
-    bad_request_error, circuit_breaker.
-    """
-    dispatch = _PROVIDER_DISPATCH.get(provider)
-    if dispatch is None:
-        raise ValueError(f"Unsupported provider: {provider}")
-    factory = dispatch.get(key)
-    if factory is None:
-        raise KeyError(f"Unknown config key '{key}' for provider '{provider}'")
-    return factory()
-
-
-# Convenience wrappers (preserve existing call sites)
-def _get_provider_client(provider: str) -> Any:
-    return _get_provider_config(provider, "get_client")
-
-
-def _get_fallback_chain(provider: str) -> list[str]:
-    result: list[str] = _get_provider_config(provider, "fallback_chain")
-    return result
-
-
-def _get_provider_retryable_errors(provider: str) -> tuple[type[Exception], ...]:
-    result: tuple[type[Exception], ...] = _get_provider_config(provider, "retryable_errors")
-    return result
-
-
-def _get_provider_bad_request_error(provider: str) -> type[Exception] | None:
-    result: type[Exception] | None = _get_provider_config(provider, "bad_request_error")
-    return result
-
-
-def _get_provider_circuit_breaker(provider: str) -> CircuitBreaker:
-    result: CircuitBreaker = _get_provider_config(provider, "circuit_breaker")
-    return result
-
-
-def _retry_provider_aware(
-    fn: Any,
-    *,
-    model: str,
-    provider: str,
-) -> Any:
-    """Execute fn with retry + backoff + fallback, provider-aware."""
-    fallback = _get_fallback_chain(provider)
-    candidates = [model] + [m for m in fallback if m != model]
-    models_to_try = [m for m in candidates if is_model_allowed(m)]
-    if not models_to_try:
-        raise RuntimeError(f"All models blocked by policy: {candidates}")
-
-    return retry_with_backoff_generic(
-        fn,
-        model=models_to_try[0],
-        fallback_models=models_to_try[1:],
-        circuit_breaker=_get_provider_circuit_breaker(provider),
-        retryable_errors=_get_provider_retryable_errors(provider),
-        bad_request_error=_get_provider_bad_request_error(provider),
-        billing_message=f"{provider} API billing/credit error.",
-        provider_label=provider.upper(),
-    )
-
-
-T_Result = TypeVar("T_Result")
-
-
-def _cross_provider_dispatch(  # noqa: UP047 — PEP695 syntax requires Python 3.12+
-    primary_provider: str,
-    primary_model: str,
-    dispatch_fn: Callable[[str, str], T_Result],
-    function_name: str,
-) -> T_Result:
-    """Execute dispatch_fn(provider, model) with opt-in cross-provider fallback.
-
-    When ``settings.llm_cross_provider_failover`` is True and the primary
-    provider chain is exhausted, iterates through ``llm_cross_provider_order``
-    trying each remaining provider's primary model.
-    """
-    providers: list[tuple[str, str]] = [(primary_provider, primary_model)]
-    if settings.llm_cross_provider_failover:
-        for p in settings.llm_cross_provider_order:
-            if p != primary_provider:
-                chain = _get_fallback_chain(p)
-                if chain:
-                    providers.append((p, chain[0]))
-
-    last_exc: Exception | None = None
-    t0 = time.perf_counter()
-    for idx, (provider, model) in enumerate(providers):
-        try:
-            return dispatch_fn(provider, model)
-        except Exception as exc:
-            last_exc = exc
-            if idx < len(providers) - 1:
-                elapsed_ms = (time.perf_counter() - t0) * 1000
-                next_p, next_m = providers[idx + 1]
-                log.warning(
-                    "Cross-provider fallback: %s(%s) -> %s(%s) [%s] after %.0fms",
-                    provider,
-                    model,
-                    next_p,
-                    next_m,
-                    function_name,
-                    elapsed_ms,
-                )
-                _fire_hook(
-                    "fallback_cross_provider",
-                    {
-                        "from_provider": provider,
-                        "to_provider": next_p,
-                        "from_model": model,
-                        "to_model": next_m,
-                        "function": function_name,
-                        "error": str(exc),
-                        "elapsed_ms": round(elapsed_ms, 1),
-                        "attempt": idx,
-                    },
-                )
-                continue
-            raise
-
-    # Unreachable when providers is non-empty; satisfies type checker
-    assert last_exc is not None
-    raise last_exc
 
 
 # ---------------------------------------------------------------------------
@@ -1131,145 +953,6 @@ def call_llm_streaming(
 
 
 # ---------------------------------------------------------------------------
-# LLMClientPort — Protocol interface for LLM provider adapters
-# ---------------------------------------------------------------------------
-
-T2 = TypeVar("T2", bound=BaseModel)
-
-
-@runtime_checkable
-class LLMClientPort(Protocol):
-    """Protocol for LLM client adapters.
-
-    Implementations: ClaudeAdapter (router functions), OpenAIAdapter, MockAdapter.
-    """
-
-    @property
-    def model_name(self) -> str:
-        """Return the default model name for cross-LLM verification."""
-        ...
-
-    def generate(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> str: ...
-
-    def generate_structured(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> dict[str, Any]: ...
-
-    def generate_parsed(
-        self,
-        system: str,
-        user: str,
-        *,
-        output_model: type[T2],
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> T2: ...
-
-    def generate_stream(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> Iterator[str]: ...
-
-    def generate_with_tools(
-        self,
-        system: str,
-        user: str,
-        *,
-        tools: list[dict[str, Any]],
-        tool_executor: Callable[..., dict[str, Any]],
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-        max_tool_rounds: int = 5,
-    ) -> Any: ...
-
-
-# ---------------------------------------------------------------------------
-# Lightweight callable protocols for node-level DI
-# ---------------------------------------------------------------------------
-
-
-class LLMJsonCallable(Protocol):
-    """Callable that returns parsed JSON from an LLM."""
-
-    def __call__(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = ...,
-        max_tokens: int = ...,
-        temperature: float = ...,
-    ) -> dict[str, Any]: ...
-
-
-class LLMTextCallable(Protocol):
-    """Callable that returns raw text from an LLM."""
-
-    def __call__(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = ...,
-        max_tokens: int = ...,
-        temperature: float = ...,
-    ) -> str: ...
-
-
-class LLMParsedCallable(Protocol):
-    """Callable that returns a Pydantic model instance from an LLM."""
-
-    def __call__(
-        self,
-        system: str,
-        user: str,
-        *,
-        output_model: type[T2],
-        model: str | None = ...,
-        max_tokens: int = ...,
-        temperature: float = ...,
-    ) -> T2: ...
-
-
-class LLMToolCallable(Protocol):
-    """Callable that runs a tool-use loop with an LLM."""
-
-    def __call__(
-        self,
-        system: str,
-        user: str,
-        *,
-        tools: list[dict[str, Any]],
-        tool_executor: Callable[..., dict[str, Any]],
-        model: str | None = ...,
-        max_tokens: int = ...,
-        temperature: float = ...,
-        max_tool_rounds: int = ...,
-    ) -> Any: ...
-
-
-# ---------------------------------------------------------------------------
 # Thread-safe injection via contextvars
 # ---------------------------------------------------------------------------
 
@@ -1346,185 +1029,3 @@ def get_secondary_llm_json() -> LLMJsonCallable | None:
 def get_secondary_llm_parsed() -> LLMParsedCallable | None:
     """Return the secondary parsed callable, or None if not configured."""
     return _secondary_llm_parsed_ctx.get()
-
-
-# ---------------------------------------------------------------------------
-# AgenticLLMPort — Protocol interface for agentic loop LLM adapters
-# ---------------------------------------------------------------------------
-
-from core.cli.agentic_response import AgenticResponse  # noqa: E402
-
-
-@runtime_checkable
-class AgenticLLMPort(Protocol):
-    """Protocol for agentic loop LLM calls.
-
-    Implementations: ClaudeAgenticAdapter, OpenAIAgenticAdapter, GlmAgenticAdapter.
-    """
-
-    @property
-    def provider_name(self) -> str: ...
-
-    @property
-    def fallback_chain(self) -> list[str]: ...
-
-    last_error: Exception | None
-
-    async def agentic_call(
-        self,
-        *,
-        model: str,
-        system: str,
-        messages: list[dict[str, Any]],
-        tools: list[dict[str, Any]],
-        tool_choice: dict[str, str] | str,
-        max_tokens: int,
-        temperature: float,
-    ) -> AgenticResponse | None: ...
-
-    def reset_client(self) -> None: ...
-
-
-# ---------------------------------------------------------------------------
-# resolve_agentic_adapter — factory + cross-provider fallback map
-# ---------------------------------------------------------------------------
-
-import importlib  # noqa: E402
-
-from core.config import ANTHROPIC_PRIMARY, OPENAI_PRIMARY  # noqa: E402
-
-# Provider -> "module_path:ClassName"
-_ADAPTER_MAP: dict[str, str] = {
-    "anthropic": "core.llm.providers.anthropic:ClaudeAgenticAdapter",
-    "openai": "core.llm.providers.openai:OpenAIAgenticAdapter",
-    "glm": "core.llm.providers.glm:GlmAgenticAdapter",
-}
-
-# Cross-provider fallback: when a provider's chain is exhausted, try these.
-# GLM -> OpenAI -> Anthropic (Bug #6 fix: add Anthropic path for GLM)
-CROSS_PROVIDER_FALLBACK: dict[str, list[tuple[str, str]]] = {
-    "anthropic": [("openai", OPENAI_PRIMARY)],
-    "openai": [("anthropic", ANTHROPIC_PRIMARY)],
-    "glm": [("openai", OPENAI_PRIMARY), ("anthropic", ANTHROPIC_PRIMARY)],
-}
-
-_router_log = logging.getLogger(__name__)
-
-
-def resolve_agentic_adapter(provider: str) -> AgenticLLMPort:
-    """Create an agentic adapter for the given provider.
-
-    Uses dynamic import to avoid loading unused providers.
-    """
-    entry = _ADAPTER_MAP.get(provider)
-    if entry is None:
-        # Unknown provider -> default to OpenAI-compatible
-        _router_log.warning("Unknown provider '%s', defaulting to openai adapter", provider)
-        entry = _ADAPTER_MAP["openai"]
-
-    module_path, class_name = entry.rsplit(":", 1)
-    mod = importlib.import_module(module_path)
-    cls = getattr(mod, class_name)
-    adapter: AgenticLLMPort = cls()
-    return adapter
-
-
-# ---------------------------------------------------------------------------
-# ClaudeAdapter — thin wrapper that delegates to router functions
-# ---------------------------------------------------------------------------
-
-
-class ClaudeAdapter:
-    """Anthropic Claude adapter implementing LLMClientPort.
-
-    Wraps the router functions into the port interface.
-    """
-
-    @property
-    def model_name(self) -> str:
-        """Return the default model name for cross-LLM verification."""
-        return settings.model
-
-    def generate(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> str:
-        result: str = call_llm(
-            system, user, model=model, max_tokens=max_tokens, temperature=temperature
-        )
-        return result
-
-    def generate_structured(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> dict[str, Any]:
-        result: dict[str, Any] = call_llm_json(
-            system, user, model=model, max_tokens=max_tokens, temperature=temperature
-        )
-        return result
-
-    def generate_parsed(
-        self,
-        system: str,
-        user: str,
-        *,
-        output_model: type[T],
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> T:
-        return call_llm_parsed(  # type: ignore[no-any-return]
-            system,
-            user,
-            output_model=output_model,
-            model=model,
-            max_tokens=max_tokens,
-            temperature=temperature,
-        )
-
-    def generate_stream(
-        self,
-        system: str,
-        user: str,
-        *,
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-    ) -> Iterator[str]:
-        return call_llm_streaming(  # type: ignore[no-any-return]
-            system, user, model=model, max_tokens=max_tokens, temperature=temperature
-        )
-
-    def generate_with_tools(
-        self,
-        system: str,
-        user: str,
-        *,
-        tools: list[dict[str, Any]],
-        tool_executor: Callable[..., dict[str, Any]],
-        model: str | None = None,
-        max_tokens: int = 4096,
-        temperature: float = 0.3,
-        max_tool_rounds: int = 5,
-    ) -> Any:
-        result: ToolUseResult = call_llm_with_tools(
-            system,
-            user,
-            tools=tools,
-            tool_executor=tool_executor,
-            model=model,
-            max_tokens=max_tokens,
-            temperature=temperature,
-            max_tool_rounds=max_tool_rounds,
-        )
-        return result
